ingest_python/requirements.txt
# Core Dependencies
PyMuPDF==1.26.7
openai==1.109.1
asyncpg==0.31.0
SQLAlchemy==2.0.44
aiohttp==3.13.2
nltk==3.9.2
tiktoken==0.12.0
python-dotenv==1.2.1

# Optional but recommended
psycopg2-binary==2.9.11
